+++ src/main/java/com/incentify/incentifyapi/_service/UserService.java
@@ -6,6 +6,8 @@ import java.util.Random;
 
 import com.incentify.incentifyapi._models.Tier;
 import com.incentify.incentifyapi._models.User;
+import com.incentify.incentifyapi.repository.RewardCodeRepository;
+import com.incentify.incentifyapi.repository.UserRepository;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
