+++ src/main/java/com/incentify/incentifyapi/repository/RewardRepository.java
@@ -1,4 +1,4 @@
-package com.incentify.incentifyapi._service;
+package com.incentify.incentifyapi.repository;
 
 import org.springframework.data.repository.CrudRepository;
